requirements.txt
openai
pillow
requests
